# components/onboarding/api/swagger.yaml
basePath: /
definitions:
...
  Asset:
    description: Asset represents a financial instrument within a ledger, such as
      a currency, cryptocurrency, commodity, or other asset type.
    properties:
      code:
        description: Unique code/symbol for the asset (max length 100 characters)
        example: USD
        maxLength: 100
        type: string
      createdAt:
        description: Timestamp when the asset was created
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      deletedAt:
        description: Timestamp when the asset was deleted (null if not deleted)
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
      id:
        description: Unique identifier for the asset (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      ledgerId:
        description: ID of the ledger this asset belongs to (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      metadata:
        additionalProperties: {}
        description: Additional custom attributes for the asset
        type: object
      name:
        description: Name of the asset (max length 256 characters)
        example: US Dollar
        maxLength: 256
        type: string
      organizationId:
        description: ID of the organization that owns this asset (UUID format)
        example: 00000000-0000-0000-0000-000000000000
        format: uuid
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Status of the asset (active, inactive, pending)
      type:
        description: Type of the asset (e.g., currency, cryptocurrency, commodity,
          stock)
        example: currency
        type: string
      updatedAt:
        description: Timestamp when the asset was last updated
        example: "2021-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
  CreateAssetInput:
    description: CreateAssetInput is the input payload to create an asset within a
      ledger, such as a currency, cryptocurrency, or other financial instrument.
    properties:
      code:
        description: Unique code/symbol for the asset (required, max length 100 characters)
        example: USD
        maxLength: 100
        type: string
      metadata:
        additionalProperties: {}
        description: |-
          Additional custom attributes for the asset
          Keys max length: 100 characters, Values max length: 2000 characters
        type: object
      name:
        description: Name of the asset (required, max length 256 characters)
        example: US Dollar
        maxLength: 256
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Status of the asset (active, inactive, pending)
      type:
        description: Type of the asset (e.g., currency, cryptocurrency, commodity,
          stock)
        example: currency
        type: string
    required:
    - code
    - name
    - type
    type: object
...
  Error:
    description: Standardized error response format used across all API endpoints
      for error situations. Provides structured information about errors including
      codes, messages, and field-specific validation details.
    properties:
      code:
        description: |-
          Error code identifying the specific error condition
          example: ERR_INVALID_INPUT
          maxLength: 50
        example: ERR_INVALID_INPUT
        maxLength: 50
        type: string
      entityType:
        description: |-
          Optional type of entity associated with the error
          example: Organization
          maxLength: 100
        example: Organization
        maxLength: 100
        type: string
      fields:
        additionalProperties:
          type: string
        description: |-
          Optional detailed field validations for client-side handling
          example: {"name": "Field 'name' is required"}
        example:
          '{"name"': ' "Field ''name'' is required"}'
        type: object
      message:
        description: |-
          Detailed error message explaining the issue
          example: The request contains invalid fields. Please check the field 'name' and try again.
          maxLength: 500
        example: The request contains invalid fields. Please check the field 'name'
          and try again.
        maxLength: 500
        type: string
      title:
        description: |-
          Short, human-readable error title
          example: Bad Request
          maxLength: 100
        example: Bad Request
        maxLength: 100
        type: string
    type: object
...
  Pagination:
    description: Pagination is the struct designed to store the pagination data of
      an entity list.
    properties:
      items: {}
      limit:
        example: 10
        type: integer
      next_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
      page:
        example: 1
        type: integer
      prev_cursor:
        example: MDAwMDAwMDAtMDAwMC0wMDAwLTAwMDAtMDAwMDAwMDAwMDAwMA==
        type: string
        x-omitempty: true
    type: object
...
  Status:
    description: 'Entity status information with a standardized code and optional
      description. Common status codes include: ACTIVE, INACTIVE, PENDING, SUSPENDED,
      DELETED.'
    properties:
      code:
        description: 'Status code identifier, common values include: ACTIVE, INACTIVE,
          PENDING, SUSPENDED, DELETED'
        example: ACTIVE
        maxLength: 100
        type: string
      description:
        description: Optional human-readable description of the status
        example: Active status
        maxLength: 256
        type: string
    type: object
...
  UpdateAssetInput:
    description: UpdateAssetInput is the input payload to update an existing asset's
      properties such as name, status, and metadata.
    properties:
      metadata:
        additionalProperties: {}
        description: |-
          Updated or additional custom attributes for the asset
          Keys max length: 100 characters, Values max length: 2000 characters
        type: object
      name:
        description: Updated name of the asset (optional, max length 256 characters)
        example: Bitcoin
        maxLength: 256
        type: string
      status:
        allOf:
        - $ref: '#/definitions/Status'
        description: Updated status of the asset (active, inactive, pending)
    type: object
...
host: localhost:3000
info:
  contact:
    name: Discord community
    url: https://discord.gg/DnhqKwkGv3
  description: This is a swagger documentation for the Midaz Ledger API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Midaz Onboarding API
  version: v1.48.0
paths:
...
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/assets:
    get:
      description: Returns a paginated list of assets within the specified ledger,
        optionally filtered by metadata, date range, and other criteria
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: JSON string to filter assets by metadata fields
        in: query
        name: metadata
        type: string
      - default: 10
        description: Maximum number of records to return per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: Page number for pagination
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Filter assets created on or after this date (format: YYYY-MM-DD)'
        in: query
        name: start_date
        type: string
      - description: 'Filter assets created on or before this date (format: YYYY-MM-DD)'
        in: query
        name: end_date
        type: string
      - description: Sort direction for results based on creation date
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved assets list
          schema:
            allOf:
            - $ref: '#/definitions/Pagination'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/Asset'
                  type: array
                limit:
                  type: integer
                page:
                  type: integer
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization or ledger not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: List all assets
      tags:
      - Assets
    post:
      consumes:
      - application/json
      description: Creates a new asset within the specified ledger. Assets represent
        currencies, cryptocurrencies, commodities, or other financial instruments
        tracked in the ledger.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Asset details including name, code, type, status, and optional
          metadata
        in: body
        name: asset
        required: true
        schema:
          $ref: '#/definitions/CreateAssetInput'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created asset
          schema:
            $ref: '#/definitions/Asset'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Organization or ledger not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Asset with the same name or code already exists'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Create a new asset
      tags:
      - Assets
  /v1/organizations/{organization_id}/ledgers/{ledger_id}/assets/{id}:
    delete:
      description: Permanently removes an asset from the specified ledger. This operation
        cannot be undone.
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Asset ID in UUID format
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Asset successfully deleted
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Asset, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Asset cannot be deleted due to existing dependencies'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Delete an asset
      tags:
      - Assets
    get:
      description: Returns detailed information about an asset identified by its UUID
        within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Asset ID in UUID format
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved asset
          schema:
            $ref: '#/definitions/Asset'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Asset, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Retrieve a specific asset
      tags:
      - Assets
    patch:
      consumes:
      - application/json
      description: Updates an existing asset's properties such as name, status, and
        metadata within the specified ledger
      parameters:
      - description: 'Authorization Bearer Token with format: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request ID for tracing
        in: header
        name: X-Request-Id
        type: string
      - description: Organization ID in UUID format
        in: path
        name: organization_id
        required: true
        type: string
      - description: Ledger ID in UUID format
        in: path
        name: ledger_id
        required: true
        type: string
      - description: Asset ID in UUID format
        in: path
        name: id
        required: true
        type: string
      - description: Asset properties to update including name, status, and optional
          metadata
        in: body
        name: asset
        required: true
        schema:
          $ref: '#/definitions/UpdateAssetInput'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated asset
          schema:
            $ref: '#/definitions/Asset'
        "400":
          description: Invalid input, validation errors
          schema:
            $ref: '#/definitions/Error'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/Error'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/Error'
        "404":
          description: Asset, ledger, or organization not found
          schema:
            $ref: '#/definitions/Error'
        "409":
          description: 'Conflict: Asset with the same name already exists'
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/Error'
      summary: Update an asset
      tags:
      - Assets
...
swagger: "2.0"